dropboxd will now be run from inside the $DROPBOX_DATA folder
One known limitation of using Dropbox for the first time in a
headless environment is that it will delete any folder named
Dropbox when registering for the first time. This deletes the
symlink I create on initial start.

Since I can't find a good workaround to this limitation, the
best option for me now is to do the reverse: symlink the
.dropbox-dist folder into the DROPBOX_DATA folder, and run
dropboxd from there. I am not sure what are the implications
of this when using data volume containers, but for now this
is the best option I have.